+++ src/main/java/model/ladder/line/Line.java
@@ -1,4 +1,4 @@
-package model.line;
+package model.ladder.line;
 
 import java.util.List;
 import model.Index;
