+++ src/test/java/model/line/FixedLinesGenerator.java
@@ -2,6 +2,8 @@ package model.line;
 
 import java.util.List;
 import model.ladder.Height;
+import model.ladder.line.Line;
+import model.ladder.line.LinesGenerator;
 
 public class FixedLinesGenerator implements LinesGenerator {
     private final List<Line> lines;
