+++ src/test/java/model/line/LineTest.java
@@ -3,6 +3,8 @@ package model.line;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import model.ladder.line.Line;
+import model.ladder.line.Path;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
